Add Privacy Policy RecycleView (need fixing)

+++ data/src/main/java/com/reader/bacalagi/data/network/response/ProductResponse.kt
@@ -6,7 +6,7 @@ data class ProductResponse(
 
 	val id: String,
 	val status: String,
-	val finalPrice: Long,
+	val finalPrice: Int,
 	val recommendedPrice: Long,
 	val description: String,
 	val seenCount: Long,
